Add register post method
