+++ backend/backend/constant.py
@@ -1,0 +1,9 @@
+SUCCESS = {
+    'code': '1',
+    'msg': 'success'
+}
+FAILURE = {
+    'code': '-1',
+    'msg': 'failed',
+    'cause': ''
+}
