+++ backend/game/views.py
@@ -7,7 +7,8 @@ from rest_framework.status import (
     HTTP_204_NO_CONTENT,
     HTTP_400_BAD_REQUEST,
 )
-
+from django.http import JsonResponse
+from rest_framework.response import Response
 from rest_framework import (
     mixins,
     generics,
@@ -30,13 +31,12 @@ from rest_framework.generics import (
 )
 from rest_framework.views import APIView
 
-from game.models import Word
-from game.serializers import (
-    WordSerializer,
-    UserRegisterSerializer,
-    UserLoginSerializer,
+from backend.constant import (
+    SUCCESS,
+    FAILURE
 )
-from game.models import User
+from game.models import Word
+from game.serializers import WordSerializer
 
 
 # viewsets: provides the implementation for CRUD
@@ -46,39 +46,3 @@ class WordView(viewsets.ModelViewSet):
     ]
     serializer_class = WordSerializer
     queryset = Word.objects.all()
-
-
-class UserRegistrationAPIView(CreateAPIView):
-    model = User
-    # let anon users register
-    permission_classes = [
-        AllowAny
-    ]
-    serializer_class = UserRegisterSerializer
-
-
-class UserLoginAPIView(APIView):
-    permission_classes = [
-        AllowAny
-    ]
-    serializer_class = UserLoginSerializer
-
-    def post(self, request, *args, **kwargs):
-        data = request.data
-        serializer = UserLoginSerializer(data=data)
-        if serializer.is_valid(raise_exception=True):
-            new_data = serializer.data
-            return Response(new_data, status=HTTP_200_OK)
-        return Response(serializer.errors, status=HTTP_400_BAD_REQUEST)
-
-class UserLogoutAllView(APIView):
-    permission_classes = [IsAuthenticated]
-
-    @swagger_auto_schema(
-        operation_summary='Logout',
-    )
-    def post(self, request, *args, **kwargs):
-        user = request.user
-        user.jwt_secret = uuid.uuid4()
-        user.save()
-        return Response(status=HTTP_204_NO_CONTENT)
