+++ backend/user/urls.py
@@ -1,0 +1,32 @@
+"""backend URL Configuration
+
+The `urlpatterns` list routes URLs to views. For more information please see:
+    https://docs.djangoproject.com/en/2.2/topics/http/urls/
+Examples:
+Function views
+    1. Add an import:  from my_app import views
+    2. Add a URL to urlpatterns:  path('', views.home, name='home')
+Class-based views
+    1. Add an import:  from other_app.views import Home
+    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
+Including another URLconf
+    1. Import the include() function: from django.urls import include, path
+    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
+"""
+from django.contrib import admin
+from django.urls import (
+    path,
+    include,
+    re_path
+)
+
+from rest_framework_jwt.views import obtain_jwt_token, refresh_jwt_token
+from user import views
+
+
+urlpatterns = [
+    path('login/', obtain_jwt_token),
+    path('register/', views.UserRegistrationAPIView.as_view()),
+    path('refresh_token/', refresh_jwt_token),
+    path('logout/', views.UserLogoutAllView.as_view()),
+]
