+++ frontend/package.json
@@ -8,6 +8,8 @@
     "axios": "^0.19.0",
     "bootstrap": "^4.3.1",
     "customize-cra": "^0.8.0",
+    "es-abstract": "^1.16.2",
+    "http-proxy-middleware": "^0.20.0",
     "react": "^16.8.6",
     "react-app-rewired": "^2.1.3",
     "react-bootstrap": "^1.0.0-beta.14",
