+++ frontend/src/components/Const.js
@@ -3,3 +3,5 @@ export const FAMILIAR = 0;
 export const REMEMBERED = 1;
 export const REGISTER_SUCCESS = 'REGISTER_SUCCESS';
 export const REGISTER_FAILURE = 'REGISTER_FAILURE';
+export const SUCCESS = '1';
+export const FAILURE = '-1';
