+++ frontend/src/redux/user.redux.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import {REGISTER_SUCCESS, REGISTER_FAILURE} from './../components/Const';
+import {REGISTER_SUCCESS, REGISTER_FAILURE, SUCCESS} from './../components/Const';
 
 /**
  * 這是一個 reducer，一個有 (state, action) => state signature 的 pure function。
@@ -28,6 +28,7 @@ let initState = {
 export function user(state=initState, action) {
     switch (action.type) {
        case REGISTER_SUCCESS:
+            alert('####', 'login!')
             return {...state, ...action.data, msg: '', redirectTo: '/login'}
        case REGISTER_FAILURE:
             return {...state, msg: action.msg}
@@ -62,13 +63,24 @@ export function register({email, username, password, pwdConfirm}) {
         return registerFail('兩次密碼不一致')
     }
     return dispatch => {
-        axios.post('/user/register/',{email, username, password})
+        axios.post('/user/register/',{
+            'email': email,
+            'username': username,
+            'password': password
+        })
             .then(res => {
-                if(res.status === 200 && res.data.code === 0) {
-                    dispatch(registerSuccess(res.data.data))
-                }else {
-                    dispatch(registerFail(res.data.msg))
-                }
+                alert('123');
+                console.log('@@@@', res.data);
+                alert('123');
+                //if(res.status === 200 && res.data.code === SUCCESS) {
+                //    console.log('####', res.data)
+                //    alert('@@@@')
+                //    dispatch(registerSuccess(res.data.data))
+                //}else {
+                //    console.log('####', res.data)
+                //    alert('@@@@@@', res)
+                //    dispatch(registerFail(res.data.msg))
+                //}
             }
         )
     }
